api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic[email]
pydantic-settings
boto3
aio_pika
websockets
passlib[bcrypt]
PyJWT
pyotp
python-multipart
email-validator
